Sendgrid error handling. for #34
