+++ lib/models/providers/email/sendgrid.js
@@ -81,6 +81,8 @@ function SendgridErrorStatus(response) {
   // There was an error when attempting to send message
   this.state.message = response.message;
 
+  // See https://sendgrid.com/docs/API_Reference/Web_API_v3/index.html for errors
+  // Currently unable to access the HTTP statusCode of the error response, so we must rely on the error messages
   if (this.state.message === 'Maximum credits exceeded') this.state.code = this.StatusCodes.LIMIT_EXCEEDED;
   else if (this.state.message === 'sendgrid error') this.state.code = this.StatusCodes.SERVICE_DOWN;
   else this.state.code = this.StatusCodes.DEFAULT;
